legacy/requirements_optimization.txt
# Minimal Requirements for PyTorch Model Optimization for Raspberry Pi Zero
# Core PyTorch (essential)
torch>=1.13.0
torchvision>=0.14.0
timm>=0.6.0

# Model optimization (essential)
onnx>=1.12.0
onnxruntime>=1.12.0

# Analysis and profiling (lightweight)
thop>=0.1.0         # For FLOPs calculation
psutil>=5.9.0       # For memory monitoring
pandas>=1.5.0       # For data analysis
tqdm>=4.64.0        # For progress bars

# Image processing (essential)
Pillow>=9.0.0       # For image loading and processing
numpy>=1.21.0       # For numerical operations

# Optional visualization (install only if needed)
# matplotlib>=3.5.0   # Commented out - install separately if needed
# seaborn>=0.11.0     # Commented out - install separately if needed

# REMOVED PACKAGES TO SAVE DISK SPACE:
# - tensorflow (200MB+) - Not needed for PyTorch optimization
# - tf2onnx - Not needed for PyTorch optimization  
# - onnxsim - Optional, can be installed separately if needed
# - torchtext - Not needed for computer vision tasks
# - matplotlib/seaborn - Optional, for visualization only

# Note: json, logging, typing, and pathlib are built-in Python modules 